+++ base/settings/tests/testthat/test.read.settings.R
@@ -15,6 +15,33 @@ testdir <- tempfile()
 dir.create(testdir, showWarnings = FALSE)
 teardown(unlink(testdir, recursive = TRUE))
 
+
+test_that("read.settings() strips comments", {
+  s_comments <- read.settings("testsettings-comment.xml")
+  s <- read.settings("testsettings.xml")
+  expect_equal(s_comments, s)
+})
+
+test_that("read.settings() warns if named input file doesn't exist (but pecan.xml does)", {
+  old_setting <- PEcAn.logger::logger.setLevel("DEBUG")
+  #hacky way to check for errors b/c PEcAn.logger errors are non-standard and
+  #not captured by testthat::expect_message() or expect_error()
+  x <- capture.output(
+    read.settings("blahblahblah.xml"),
+    type = "message"
+  )
+  expect_true(any(grepl("WARN", x)))
+  expect_true(any(grepl("blahblahblah.xml not found", x)))
+  
+  PEcAn.logger::logger.setLevel(old_setting)
+})
+
+test_that("read settings returns error if no settings file found (#1124)", {
+  withr::with_tempdir({ #in a dir with no pecan.xml
+    expect_error(read.settings("nofile.xml"), "Could not find a pecan.xml file")
+  })
+})
+
 # ANS: Many of these functions seem like they shouldn't require a
 # database connection, but `check.settings` throws an error when it
 # can't connect to the database.
@@ -29,10 +56,6 @@ test_that("read.settings returned correctly", {
   skip("Tests failing due to multisite?")
   expect_true(file.exists(s$outdir))
   expect_true(file.info(s$outdir)$isdir)
-})
-
-test_that("read settings returns error if no settings file found (#1124)", {
-  expect_error(read.settings("nofile.xml"), "Could not find a pecan.xml file")
 })